Add: Create SubscriptionTable Component
