+++ src/features/DashBoard/DashBoardPageTemplate.tsx
@@ -12,13 +12,14 @@ import { convertToMoney } from "~/utils/convertToMoney";
 import { formatUnixTimeDate } from "~/utils/formatUnixTimeDate";
 import { colors } from "styles/themes";
 import { stripeApi } from "~/context/ApiContext";
+import { SubscriptionTable } from "./components/SubscriptionTable";
 
 export const DashBoardPageTemplate = ({
   paymentList,
   subscriptionList
 }: DashBoardPageProps): JSX.Element => {
   const { currentUser, isLoading } = useRecoilValue(authState);
-  const { push, reload } = useRouter();
+  const { push } = useRouter();
 
   if (isLoading) return <Loading size="xl" />;
 
@@ -26,16 +27,6 @@ export const DashBoardPageTemplate = ({
     push("/");
     return <Fragment />;
   }
-
-  const handleClick = async (stripeCustomerId: string) => {
-    try {
-      await stripeApi.cancelSubscription(stripeCustomerId);
-
-      reload();
-    } catch (error) {
-      console.error(error);
-    }
-  };
 
   return (
     <main css={main}>
@@ -54,50 +45,7 @@ export const DashBoardPageTemplate = ({
         </FlexContainer>
         <div>
           <p css={label}>サブスク一覧</p>
-          <Table
-            aria-label="Example static collection table"
-            css={{
-              height: "auto",
-              fontSize: "1.4rem"
-            }}
-          >
-            <Table.Header>
-              <Table.Column>金額</Table.Column>
-              <Table.Column>ステータス</Table.Column>
-              <Table.Column>契約日</Table.Column>
-              <Table.Column>契約更新日</Table.Column>
-              <Table.Column>cancel</Table.Column>
-            </Table.Header>
-            <Table.Body>
-              {subscriptionList.map((subscription) => (
-                <Table.Row key={subscription.id}>
-                  <Table.Cell>
-                    {convertToMoney(
-                      subscription.items.data[0].price.unit_amount
-                    )}
-                  </Table.Cell>
-                  <Table.Cell>{subscription.status}</Table.Cell>
-                  <Table.Cell>
-                    {formatUnixTimeDate(subscription.start_date, "jp")}
-                  </Table.Cell>
-                  <Table.Cell>
-                    {formatUnixTimeDate(subscription.current_period_end, "jp")}
-                  </Table.Cell>
-                  <Table.Cell>
-                    {subscription.status === "active" && (
-                      <Button
-                        auto
-                        color="error"
-                        onClick={() => handleClick(subscription.id)}
-                      >
-                        解約する
-                      </Button>
-                    )}
-                  </Table.Cell>
-                </Table.Row>
-              ))}
-            </Table.Body>
-          </Table>
+          <SubscriptionTable subscriptionList={subscriptionList} />
         </div>
         <div style={{ marginTop: 40 }}>
           <p css={label}>注文履歴</p>
